Django/api-basic2/requirements.txt
boto3==1.9.170
botocore==1.12.170
certifi==2019.6.16
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
Django==2.2.2
django-rest-framework==0.1.0
django-rest-swagger==2.1.2
django-storages==1.7.1
djangorestframework==3.9.4
docutils==0.14
idna==2.8
itypes==1.1.0
Jinja2==2.10.1
jmespath==0.9.4
MarkupSafe==1.1.1
openapi-codec==1.3.2
Pillow==6.0.0
psycopg2-binary==2.8.3
python-dateutil==2.8.0
pytz==2019.1
requests==2.22.0
s3transfer==0.2.1
simplejson==3.16.0
six==1.12.0
sqlparse==0.3.0
uritemplate==3.0.0
urllib3==1.25.3
